+++ crates/forge/src/main.rs
@@ -2,8 +2,8 @@ use anyhow::{anyhow, bail, Context, Result};
 use camino::Utf8PathBuf;
 use clap::Parser;
 use include_dir::{include_dir, Dir};
-use scarb_metadata::MetadataCommand;
-
+use scarb_metadata::{MetadataCommand, PackageMetadata};
+use scarb_ui::args::PackagesFilter;
 use std::path::PathBuf;
 use tempfile::{tempdir, TempDir};
 
@@ -11,10 +11,12 @@ use forge::{pretty_printing, RunnerConfig};
 use forge::{run, TestFileSummary};
 
 use forge::scarb::{
-    corelib_for_package, dependencies_for_package, get_contracts_map, name_for_package,
-    paths_for_package, target_name_for_package, try_get_starknet_artifacts_path,
+    config_from_scarb_for_package, corelib_for_package, dependencies_for_package,
+    get_contracts_map, name_for_package, paths_for_package, target_dir_for_package,
+    target_name_for_package, try_get_starknet_artifacts_path,
 };
 use forge::test_case_summary::TestCaseSummary;
+use rand::{thread_rng, RngCore};
 use std::process::{Command, Stdio};
 
 mod init;
@@ -35,6 +37,27 @@ struct Args {
     /// Stop test execution after the first failed test
     #[arg(short = 'x', long)]
     exit_first: bool,
+
+    #[command(flatten)]
+    packages_filter: PackagesFilter,
+
+    /// Number of fuzzer runs
+    #[arg(short = 'r', long, value_parser = validate_fuzzer_runs_value)]
+    fuzzer_runs: Option<u32>,
+
+    /// Seed for the fuzzer
+    #[arg(short = 's', long)]
+    fuzzer_seed: Option<u64>,
+}
+
+fn validate_fuzzer_runs_value(val: &str) -> Result<u32> {
+    let parsed_val: u32 = val
+        .parse()
+        .map_err(|_| anyhow!("Failed to parse {val} as u32"))?;
+    if parsed_val < 3 {
+        bail!("Number of fuzzer runs must be greater than or equal to 3")
+    }
+    Ok(parsed_val)
 }
 
 fn load_predeployed_contracts() -> Result<TempDir> {
@@ -70,36 +93,46 @@ fn main_execution() -> Result<bool> {
 
     let scarb_metadata = MetadataCommand::new().inherit_stderr().exec()?;
 
-    let build_output = Command::new("scarb")
-        .current_dir(std::env::current_dir().context("Failed to get current directory")?)
-        .arg("build")
-        .stderr(Stdio::inherit())
-        .stdout(Stdio::inherit())
-        .output()
-        .context("Failed to build contracts with Scarb")?;
-    if !build_output.status.success() {
-        bail!("Scarb build did not succeed")
-    }
+    let packages: Vec<PackageMetadata> = args
+        .packages_filter
+        .match_many(&scarb_metadata)
+        .context("Failed to find any packages matching the specified filter")?;
+
+    let fuzzer_seed = args.fuzzer_seed.unwrap_or_else(|| thread_rng().next_u64());
 
     let mut all_failed_tests = vec![];
-    for package in &scarb_metadata.workspace.members {
-        let forge_config = forge::scarb::config_from_scarb_for_package(&scarb_metadata, package)?;
-        let (package_path, lib_path) = paths_for_package(&scarb_metadata, package)?;
-
+    for package in &packages {
+        let forge_config = config_from_scarb_for_package(&scarb_metadata, &package.id)?;
+        let (package_path, lib_path) = paths_for_package(&scarb_metadata, &package.id)?;
         std::env::set_current_dir(package_path.clone())?;
 
-        let package_name = name_for_package(&scarb_metadata, package)?;
-        let dependencies = dependencies_for_package(&scarb_metadata, package)?;
-        let target_name = target_name_for_package(&scarb_metadata, package)?;
-        let corelib_path = corelib_for_package(&scarb_metadata, package)?;
+        // TODO(#671)
+        let target_dir = target_dir_for_package(&scarb_metadata.workspace.root)?;
+
+        let build_output = Command::new("scarb")
+            .arg("build")
+            .stderr(Stdio::inherit())
+            .stdout(Stdio::inherit())
+            .output()
+            .context("Failed to build contracts with Scarb")?;
+        if !build_output.status.success() {
+            bail!("Scarb build did not succeed")
+        }
+
+        let package_name = name_for_package(&scarb_metadata, &package.id)?;
+        let dependencies = dependencies_for_package(&scarb_metadata, &package.id)?;
+        let target_name = target_name_for_package(&scarb_metadata, &package.id)?;
+        let corelib_path = corelib_for_package(&scarb_metadata, &package.id)?;
         let runner_config = RunnerConfig::new(
             args.test_filter.clone(),
             args.exact,
             args.exit_first,
+            args.fuzzer_runs,
+            args.fuzzer_seed,
             &forge_config,
         );
 
-        let contracts_path = try_get_starknet_artifacts_path(&package_path, &target_name)?;
+        let contracts_path = try_get_starknet_artifacts_path(&target_dir, &target_name)?;
         let contracts = contracts_path
             .map(|path| get_contracts_map(&path))
             .transpose()?
@@ -114,6 +147,7 @@ fn main_execution() -> Result<bool> {
             &corelib_path,
             &contracts,
             &predeployed_contracts,
+            fuzzer_seed,
         )?;
 
         let mut failed_tests = extract_failed_tests(tests_file_summaries);