+++ crates/forge/tests/e2e/mod.rs
@@ -1,4 +1,7 @@
-mod collection;
 pub(crate) mod common;
+
+mod collection;
 mod file_operations;
+mod fuzzing;
 mod running;
+mod workspaces;