+++ CHANGELOG.md
@@ -5,12 +5,38 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-# [Unreleased]
-# [0.6.0] - 2023-09-13
+## [Unreleased]
 
 ### Forge
 
 #### Added
+
+- Support for scarb workspaces
+- Initial version of fuzz testing with randomly generated values
+
+### Changed
+
+- Tests are collected only from `tests` folder and a package tree
+
+### Cast
+
+#### Added
+
+- `account add` command for importing accounts to the accounts file
+- `account create` command for creating openzeppelin accounts with starkli-style keystore
+- `account deploy` command for deploying openzeppelin accounts with starkli-style keystore
+
+### Changed
+
+- `--add-profile` no longer accepts `-a` for short
+- allow the `id` property in multicalls to be referenced in the inputs of `deploy` and `invoke` calls
+
+## [0.6.0] - 2023-09-13
+
+### Forge
+
+#### Added
+
 - `deploy_at` cheatcode
 - printing failures summary at the end of an execution
 - filtering tests now uses an absolute module tree path — it is possible to filter tests by module names, etc.